on Squid V2.0.x: I'm still looking for a report of a CARP-User how Calamaris works for him.
I removed the todo 'rewrite Mainloop'. I have build it in, and now the
Mainloop is build first and then run. But: It doesn't work as good as i hoped.
On my measurements on my (slow) 386DX40 with a very short access.log it is
slower, with a longer file (60k lines) it is faster by about 2%


git-svn-id: svn+ssh://weyoun2.vpn.cord.de/home/svn/calamaris@52 af268522-00f2-45aa-b56c-050e69c5c3bf